Split sequence view generation code into two files
